On macOS use `ioreg -w 0` instead of `ioreg -p IOUSB`
For some reason I cannot determine, `ioreg -p IOUSB` in the current
version of macOS is not listing my YubiKey when it's plugged in,
whereas just `ioreg` is. In addition, to ensure that we see the
YubiKey when it's plugged in, we should specify `-w 0` to `ioreg` so
output isn't truncated.